requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.14.0
idna==3.10
paho-mqtt==2.1.0
passlib==1.7.4
psycopg2==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
